feat: Separated model-endpoints, better error-handling and input-validation #31
